@@ -55,9 +55,7 @@ class Users extends BaseController
     {
         return view('users/index', [
             'title' => 'Staff',
-            'rows'  => $this->model->getCompiledRows(static function ($row) {
-                return is_numeric($row['group_id']);
-            }, 'group'),
+            'rows'  => $this->model->getCompiledRows(static fn ($row)  => is_numeric($row['group_id']), 'group'),
         ]);
     }
 
