@@ -6,7 +6,10 @@ use PhpCsFixer\Finder;
 
 $finder = Finder::create()
     ->files()
-    ->in(__DIR__)
+    ->in([
+        __DIR__ . '/app/',
+        __DIR__ . '/tests/',
+    ])
     ->exclude('build')
     ->append([__FILE__]);
 
